Introduce jvm-engine buildpack
The existing heroku/jvm buildpack both provides and requires itself unconditionally. That means if you put `heroku/jvm` in a build order it will always run even if it is specified as optional. 

What does that mean? Most build packs will always provide a resource (like "ruby") but only require it if it has a positive signal (such as presence of a Gemfile on disk). The JVM buildpack is different because the JVM ecosystem can be used with wildly different JVM applications, many of them do not have a single file that would indicate it needs the JDK. To support this set of various use cases it always installs the JVM.

This always-install behavior is a problem for the Ruby buildpack due to JRuby support. JRuby is provided by the same buildpack as Ruby the only difference is that it needs the JVM installed (and it downloads a jruby binary from S3). To install the JVM the heroku/ruby builder has `heroku/jvm` specified as an optional dependency. Due to existing behavior that means the JVM is installed for ALL ruby applications even if is not using the JVM. This eats up disk space and takes up build time.

There are possible fixes:

- A: Introduce an only-provide buildpack and change the existing `heroku/jvm` buildpack to be a no-op that always requires the other buildpack. Then the `heroku/ruby` (and any other build packs) can optionally depend on the JVM through the regular require interface.
- B: Introduce a new JRuby buildpack that fails to detect if jruby is not detected (via inspecting the Gemfile.lock). Use this to make a new group along with the heroku/jvm and place that group above the existing heroku/ruby group. Remove the `heroku/jvm` group from the heroku/ruby group. This would have the effect of only installing the JVM for JRuby apps.
- C: Some other option

This PR implements option A. It moves the existing logic to a `heroku/jvm-engine` buildpack and removes its `require`. Then the `heroku/jvm` buildpack is stripped of all logic and all it does is `require` the jvm provided by the `heroku/jvm-engine`.

The upside of this approach is that any buildpack can use the `heroku-jvm-engine` much like the `heroku-nodejs-engine` to install the JVM optionally. As long as the existing `heroku/jvm` builder is updated with both build packs then this should be a non-breaking change save if someone was using the `heroku/jvm` buildpack directly (not through the builder), they'll now need to also add the `heroku/jvm-engine` buildpack as well.


## Other thoughts

This interface is really funky. It's a very-round about way of expressing what should be a seemingly common use case. Having the builder separated from the buildpack and provide/require be based on a string instead of a specific resource leads to very difficult to reason about resolution logic. 

Another option (but it won't meet our timescale) could be a change to the builder spec: For instance we could expose the `optional` flag to a buildpack in the detect phase to let a buildpack know if it is being treated as optional or not and change its behavior accordingly.

Longer term it's worth experimenting with alternative interfaces that we could build on top of these three primitives (builder, buildpack, provide/require build plans). I've proposed a theoretical middleware like interface (Rack middleware would be my design inspiration as I come from Ruby) but I've not sketched up any kind of implementation ideas.
