+++ buildpacks/jvm-engine/buildpack.toml
@@ -1,0 +1,26 @@
+api = "0.9"
+
+[buildpack]
+id = "heroku/jvm-engine"
+version = "1.1.1"
+name = "Heroku OpenJDK Buildpack"
+homepage = "https://github.com/heroku/buildpacks-jvm"
+description = "Official Heroku buildpack for installing OpenJDK"
+keywords = ["java", "jvm", "jdk", "openjdk"]
+
+[[buildpack.licenses]]
+type = "BSD-3-Clause"
+
+[[stacks]]
+id = "heroku-20"
+
+[[stacks]]
+id = "heroku-22"
+
+[metadata]
+[metadata.heroku-metrics-agent]
+url = "https://repo1.maven.org/maven2/com/heroku/agent/heroku-java-metrics-agent/4.0.1/heroku-java-metrics-agent-4.0.1.jar"
+sha256 = "9a718680e4a93f93a8755b20fb21cc541e5c2692acc9da27c667530f48a716db"
+[metadata.release]
+[metadata.release.docker]
+repository = "docker.io/heroku/buildpack-jvm"
