+++ buildpacks/jvm-engine/package.toml
@@ -1,0 +1,2 @@
+[buildpack]
+uri = "."
