+++ buildpacks/jvm/Cargo.toml
@@ -6,16 +6,6 @@ edition.workspace = true
 publish.workspace = true
 
 [dependencies]
-buildpacks-jvm-shared.workspace = true
-fs_extra = "1"
-indoc = "2"
 libcnb.workspace = true
-libherokubuildpack.workspace = true
-serde = { version = "1", features = ["derive"] }
-tempfile = "3"
-ureq = "2"
-url = "2"
 
 [dev-dependencies]
-libcnb-test.workspace = true
-buildpacks-jvm-shared-test.workspace = true
